+++ src/tm_devices/helpers/logging.py
@@ -90,9 +90,26 @@ class LoggingLevels(CustomStrEnum):
 
 
 @contextmanager
-def disable_all_loggers() -> Generator[None, None, None]:
-    """Temporarily disable all logging within this context manager block."""
-    logging.disable(logging.CRITICAL)
+def disable_all_loggers(
+    level: Literal["DEBUG", "INFO", "WARNING", "ERROR", "CRITICAL"] = "CRITICAL",
+) -> Generator[None, None, None]:
+    """Temporarily disable all logging calls of severity 'level' and below within this context.
+
+    Examples:
+        >>> with disable_all_loggers():
+        ...     logging.info("This is inside the context and will not be logged")
+
+        >>> logging.info("This is outside the context and will be logged")
+
+        >>> with disable_all_loggers(level="WARNING"):
+        ...     logging.info("This will not be logged")
+        ...     logging.warning("This will not be logged")
+        ...     logging.error("This will be logged since its level is above WARNING")
+
+    Args:
+        level: The logging level under which all log messages will be suppressed.
+    """
+    logging.disable(getattr(logging, level))
     yield
     logging.disable(logging.NOTSET)